ui fix + general refactor+cleanup

+++ src/main/java/io/github/alerithe/client/ui/click/elements/impl/Label.java
@@ -3,7 +3,7 @@ package io.github.alerithe.client.ui.click.elements.impl;
 import io.github.alerithe.client.ui.click.elements.Element;
 
 public class Label extends Element {
-    public Label(String text, int x, int y, int width, int height) {
+    public Label(String text, float x, float y, float width, float height) {
         super(x, y, width, height);
 
         setText(text);
